camera/requirements.txt
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
python-dotenv==1.0.1
setuptools==69.5.1
